+++ components/Icon/index.ts
@@ -1,4 +1,5 @@
 export {
   ExternalLinkIcon,
-  NoteIcon
+  NoteIcon,
+  UnreadIndicator
 } from './Icon'